chore: add Knip to CI, fix circular deps, and trim unused exports
Follow-up to the lint config audit — completes remaining audit items:
Knip integration, circular dependency resolution, and unused export cleanup.

**Knip CI integration:**
- Install knip as workspace devDependency
- Add `knip.json` config with exclusions for Prisma generated code, standalone
  scripts, dynamic deps (pino-roll), and intentional API surface (React Query
  hooks, design system tokens, component libraries)
- Add `pnpm exec knip --no-exit-code` step to CI workflow (informational —
  Knip has false positives on Fastify plugin classes and Prisma re-exports)

**Circular dependencies fixed (4 → 0 real cycles):**
- API: Extract shared `Logger` type from validate-batch.ts → types.ts to break
  validate-batch ↔ schema-fingerprint cycle
- Web: Extract `InstanceUrlMap` type from dashboard-client.tsx → types.ts to
  break dashboard-client ↔ queue-table ↔ queue-group-card cycles
- Web: Move `CFResolutionDecision` and `ResolvedCF` types from cf-resolution.tsx
  → cf-configuration-types.ts to break cf-resolution ↔ cf-resolution-items cycle

**Unused exports removed (~60 across API + Web):**
- API: Service-specific client wrappers (10 functions), error type guards (4),
  unused API path helpers, hunting/queue-cleaner constants, sync executor
  functions, default export duplicates, unused types
- Web: Dead utility functions, deprecated hook aliases, unused API client
  functions, dead type exports, stale re-exports from barrel files
