+++ apps/api/src/lib/arr/client-helpers.ts
@@ -203,91 +203,6 @@ export async function executeOnInstances<T>(
 	};
 }
 
-/**
- * Execute an operation on Sonarr instances only
- */
-export async function executeOnSonarrInstances<T>(
-	app: FastifyInstance,
-	userId: string,
-	operation: (client: SonarrClient, instance: ServiceInstance) => Promise<T>,
-	options?: Omit<MultiInstanceOptions, "serviceTypes">,
-): Promise<MultiInstanceResponse<T>> {
-	return executeOnInstances(
-		app,
-		userId,
-		{ ...options, serviceTypes: ["SONARR"] },
-		(client, instance) => operation(client as SonarrClient, instance),
-	);
-}
-
-/**
- * Execute an operation on Radarr instances only
- */
-export async function executeOnRadarrInstances<T>(
-	app: FastifyInstance,
-	userId: string,
-	operation: (client: RadarrClient, instance: ServiceInstance) => Promise<T>,
-	options?: Omit<MultiInstanceOptions, "serviceTypes">,
-): Promise<MultiInstanceResponse<T>> {
-	return executeOnInstances(
-		app,
-		userId,
-		{ ...options, serviceTypes: ["RADARR"] },
-		(client, instance) => operation(client as RadarrClient, instance),
-	);
-}
-
-/**
- * Execute an operation on Prowlarr instances only
- */
-export async function executeOnProwlarrInstances<T>(
-	app: FastifyInstance,
-	userId: string,
-	operation: (client: ProwlarrClient, instance: ServiceInstance) => Promise<T>,
-	options?: Omit<MultiInstanceOptions, "serviceTypes">,
-): Promise<MultiInstanceResponse<T>> {
-	return executeOnInstances(
-		app,
-		userId,
-		{ ...options, serviceTypes: ["PROWLARR"] },
-		(client, instance) => operation(client as ProwlarrClient, instance),
-	);
-}
-
-/**
- * Execute an operation on Lidarr instances only
- */
-export async function executeOnLidarrInstances<T>(
-	app: FastifyInstance,
-	userId: string,
-	operation: (client: LidarrClient, instance: ServiceInstance) => Promise<T>,
-	options?: Omit<MultiInstanceOptions, "serviceTypes">,
-): Promise<MultiInstanceResponse<T>> {
-	return executeOnInstances(
-		app,
-		userId,
-		{ ...options, serviceTypes: ["LIDARR"] },
-		(client, instance) => operation(client as LidarrClient, instance),
-	);
-}
-
-/**
- * Execute an operation on Readarr instances only
- */
-export async function executeOnReadarrInstances<T>(
-	app: FastifyInstance,
-	userId: string,
-	operation: (client: ReadarrClient, instance: ServiceInstance) => Promise<T>,
-	options?: Omit<MultiInstanceOptions, "serviceTypes">,
-): Promise<MultiInstanceResponse<T>> {
-	return executeOnInstances(
-		app,
-		userId,
-		{ ...options, serviceTypes: ["READARR"] },
-		(client, instance) => operation(client as ReadarrClient, instance),
-	);
-}
-
 // ============================================================================
 // Single Instance Operations
 // ============================================================================
@@ -334,166 +249,6 @@ export async function getClientForInstance(
 		}
 		throw error;
 	}
-}
-
-/**
- * Get a typed Sonarr client for a specific instance
- */
-export async function getSonarrClientForInstance(
-	app: FastifyInstance,
-	request: FastifyRequest,
-	instanceId: string,
-): Promise<
-	| { success: true; client: SonarrClient; instance: ServiceInstance }
-	| { success: false; error: string; statusCode: number }
-> {
-	const result = await getClientForInstance(app, request, instanceId);
-
-	if (!result.success) {
-		return result;
-	}
-
-	if (result.instance.service !== "SONARR") {
-		return {
-			success: false,
-			error: "Instance is not a Sonarr instance",
-			statusCode: 400,
-		};
-	}
-
-	return {
-		success: true,
-		client: result.client as SonarrClient,
-		instance: result.instance,
-	};
-}
-
-/**
- * Get a typed Radarr client for a specific instance
- */
-export async function getRadarrClientForInstance(
-	app: FastifyInstance,
-	request: FastifyRequest,
-	instanceId: string,
-): Promise<
-	| { success: true; client: RadarrClient; instance: ServiceInstance }
-	| { success: false; error: string; statusCode: number }
-> {
-	const result = await getClientForInstance(app, request, instanceId);
-
-	if (!result.success) {
-		return result;
-	}
-
-	if (result.instance.service !== "RADARR") {
-		return {
-			success: false,
-			error: "Instance is not a Radarr instance",
-			statusCode: 400,
-		};
-	}
-
-	return {
-		success: true,
-		client: result.client as RadarrClient,
-		instance: result.instance,
-	};
-}
-
-/**
- * Get a typed Prowlarr client for a specific instance
- */
-export async function getProwlarrClientForInstance(
-	app: FastifyInstance,
-	request: FastifyRequest,
-	instanceId: string,
-): Promise<
-	| { success: true; client: ProwlarrClient; instance: ServiceInstance }
-	| { success: false; error: string; statusCode: number }
-> {
-	const result = await getClientForInstance(app, request, instanceId);
-
-	if (!result.success) {
-		return result;
-	}
-
-	if (result.instance.service !== "PROWLARR") {
-		return {
-			success: false,
-			error: "Instance is not a Prowlarr instance",
-			statusCode: 400,
-		};
-	}
-
-	return {
-		success: true,
-		client: result.client as ProwlarrClient,
-		instance: result.instance,
-	};
-}
-
-/**
- * Get a typed Lidarr client for a specific instance
- */
-export async function getLidarrClientForInstance(
-	app: FastifyInstance,
-	request: FastifyRequest,
-	instanceId: string,
-): Promise<
-	| { success: true; client: LidarrClient; instance: ServiceInstance }
-	| { success: false; error: string; statusCode: number }
-> {
-	const result = await getClientForInstance(app, request, instanceId);
-
-	if (!result.success) {
-		return result;
-	}
-
-	if (result.instance.service !== "LIDARR") {
-		return {
-			success: false,
-			error: "Instance is not a Lidarr instance",
-			statusCode: 400,
-		};
-	}
-
-	return {
-		success: true,
-		client: result.client as LidarrClient,
-		instance: result.instance,
-	};
-}
-
-/**
- * Get a typed Readarr client for a specific instance
- */
-export async function getReadarrClientForInstance(
-	app: FastifyInstance,
-	request: FastifyRequest,
-	instanceId: string,
-): Promise<
-	| { success: true; client: ReadarrClient; instance: ServiceInstance }
-	| { success: false; error: string; statusCode: number }
-> {
-	const result = await getClientForInstance(app, request, instanceId);
-
-	if (!result.success) {
-		return result;
-	}
-
-	if (result.instance.service !== "READARR") {
-		return {
-			success: false,
-			error: "Instance is not a Readarr instance",
-			statusCode: 400,
-		};
-	}
-
-	return {
-		success: true,
-		client: result.client as ReadarrClient,
-		instance: result.instance,
-	};
 }
 
 // ============================================================================
