+++ apps/api/src/lib/library-sync/index.ts
@@ -5,11 +5,9 @@
  */
 
 export {
-	removeCachedItem,
 	type SyncExecutorDeps,
 	type SyncResult,
 	syncInstance,
-	syncSingleItem,
 } from "./sync-executor.js";
 
 export { getLibrarySyncScheduler } from "./sync-scheduler.js";
