+++ apps/api/src/lib/library-sync/sync-executor.ts
@@ -333,70 +333,3 @@ export async function syncInstance(
 	return result;
 }
 
-/**
- * Syncs a single item to the cache (used for optimistic updates)
- */
-export async function syncSingleItem(
-	deps: Pick<SyncExecutorDeps, "prisma" | "log">,
-	instanceId: string,
-	item: LibraryItem,
-): Promise<void> {
-	const { prisma, log } = deps;
-	const arrItemId = typeof item.id === "string" ? Number.parseInt(item.id, 10) : item.id;
-	const fields = extractCacheFields(item);
-
-	try {
-		await prisma.libraryCache.upsert({
-			where: {
-				instanceId_arrItemId_itemType: {
-					instanceId,
-					arrItemId,
-					itemType: item.type,
-				},
-			},
-			create: buildCacheEntry(instanceId, item),
-			update: {
-				...fields,
-				data: JSON.stringify(item),
-				updatedAt: new Date(),
-			},
-		});
-
-		log.debug({ instanceId, arrItemId, title: item.title }, "Single item synced to cache");
-	} catch (error) {
-		log.error({ err: error, instanceId, arrItemId }, "Failed to sync single item to cache");
-		throw error;
-	}
-}
-
-/**
- * Removes a single item from the cache
- */
-export async function removeCachedItem(
-	deps: Pick<SyncExecutorDeps, "prisma" | "log">,
-	instanceId: string,
-	arrItemId: number,
-	itemType: "movie" | "series" | "artist" | "author",
-): Promise<void> {
-	const { prisma, log } = deps;
-
-	try {
-		await prisma.libraryCache.delete({
-			where: {
-				instanceId_arrItemId_itemType: {
-					instanceId,
-					arrItemId,
-					itemType,
-				},
-			},
-		});
-
-		log.debug({ instanceId, arrItemId, itemType }, "Item removed from cache");
-	} catch (error) {
-		// Item might not exist, which is fine
-		if ((error as { code?: string }).code !== "P2025") {
-			log.error({ err: error, instanceId, arrItemId }, "Failed to remove item from cache");
-			throw error;
-		}
-	}
-}
