+++ apps/api/src/routes/manual-import-utils.ts
@@ -284,75 +284,6 @@ const mapCandidate = (
 	}
 };
 
-export const fetchManualImportCandidates = async (
-	fetcher: (path: string, init?: RequestInit) => Promise<Response>,
-	service: ManualImportService,
-	options: ManualImportFetchOptions,
-): Promise<ManualImportCandidate[]> => {
-	const params = new URLSearchParams();
-
-	if (options.downloadId) {
-		params.set("downloadId", options.downloadId);
-	}
-	if (options.folder) {
-		params.set("folder", options.folder);
-	}
-	if (typeof options.seriesId === "number") {
-		params.set("seriesId", String(options.seriesId));
-	}
-	if (typeof options.seasonNumber === "number") {
-		params.set("seasonNumber", String(options.seasonNumber));
-	}
-	if (typeof options.filterExistingFiles === "boolean") {
-		params.set("filterExistingFiles", String(options.filterExistingFiles));
-	} else {
-		params.set("filterExistingFiles", "true");
-	}
-
-	const query = params.toString();
-	const apiPath = getManualImportApiPath(service);
-	const response = await fetcher(query.length > 0 ? `${apiPath}?${query}` : apiPath);
-
-	if (!response.ok) {
-		const message = await response.text().catch(() => "");
-		throw new ManualImportError(
-			`Failed to fetch manual import items (status ${response.status}): ${message || response.statusText}`,
-			response.status,
-		);
-	}
-
-	let payload: unknown;
-	try {
-		payload = await response.json();
-	} catch (_error) {
-		throw new ManualImportError("ARR returned an invalid manual import payload.");
-	}
-
-	let rawItems: unknown[];
-	if (Array.isArray(payload)) {
-		rawItems = payload;
-	} else if (Array.isArray((payload as { items?: unknown[] } | undefined)?.items)) {
-		rawItems = (payload as { items: unknown[] }).items;
-	} else {
-		// Log unexpected response structure to aid debugging API compatibility issues
-		const warnMsg =
-			`[manual-import] Unexpected API response structure for ${service}. ` +
-			`Expected array or {items: []}, got: ${typeof payload}`;
-		moduleLogger?.warn(warnMsg);
-		return [];
-	}
-
-	if (rawItems.length === 0) {
-		return [];
-	}
-
-	const candidates = rawItems
-		.map((item) => mapCandidate(service, item))
-		.filter((candidate): candidate is ManualImportCandidate => candidate !== null);
-
-	return manualImportCandidateListSchema.parse(candidates);
-};
-
 const summarizeRejections = (candidate: ManualImportCandidate): string | undefined => {
 	if (!candidate.rejections || candidate.rejections.length === 0) {
 		return undefined;
@@ -630,68 +561,6 @@ const buildCommandFiles = (
 
 		throw new ManualImportError(`Unknown service type: ${service}`);
 	});
-};
-
-export const submitManualImportCommand = async (
-	fetcher: (path: string, init?: RequestInit) => Promise<Response>,
-	service: ManualImportService,
-	files: ManualImportSubmissionFile[],
-	importMode: "auto" | "move" | "copy",
-) => {
-	if (files.length === 0) {
-		throw new ManualImportError("No files were provided for manual import.");
-	}
-
-	const commandFiles = buildCommandFiles(service, files);
-	const apiPath = getCommandApiPath(service);
-
-	const response = await fetcher(apiPath, {
-		method: "POST",
-		headers: {
-			"Content-Type": "application/json",
-		},
-		body: JSON.stringify({
-			name: "ManualImport",
-			importMode,
-			files: commandFiles,
-		}),
-	});
-
-	if (!response.ok) {
-		const message = await response.text().catch(() => "");
-		throw new ManualImportError(
-			`ARR manual import command failed (status ${response.status}): ${message || response.statusText}`,
-			response.status,
-		);
-	}
-};
-
-export const autoImportByDownloadId = async (
-	fetcher: (path: string, init?: RequestInit) => Promise<Response>,
-	service: ManualImportService,
-	downloadId: string,
-) => {
-	const candidates = await fetchManualImportCandidates(fetcher, service, {
-		downloadId,
-		filterExistingFiles: true,
-	});
-
-	if (candidates.length === 0) {
-		throw new ManualImportError("ARR did not provide any importable files for this download.");
-	}
-
-	const { files, skipped } = collectAutoManualImportFiles(service, candidates, downloadId);
-
-	if (files.length === 0) {
-		const detail = skipped.length > 0 ? skipped.slice(0, 3).join("; ") : undefined;
-		throw new ManualImportError(
-			detail
-				? `No importable files were found: ${detail}`
-				: "ARR did not provide importable files for manual import.",
-		);
-	}
-
-	await submitManualImportCommand(fetcher, service, files, "auto");
 };
 
 // ============================================================================
