+++ apps/web/src/features/dashboard/types.ts
@@ -1,0 +1,8 @@
+/**
+ * Shared types for the dashboard feature.
+ *
+ * Extracted to break the circular dependency between dashboard-client.tsx,
+ * queue-table.tsx, and queue-group-card.tsx.
+ */
+
+export type InstanceUrlMap = Map<string, string>;
