+++ apps/web/src/features/hunting/lib/constants.ts
@@ -46,22 +46,11 @@ export const MIN_HOURLY_API_CAP = 10;
 export const MAX_HOURLY_API_CAP = 500;
 
 /**
- * Default hourly API cap
- */
-export const DEFAULT_HOURLY_API_CAP = 100;
-
-/**
  * Maximum queue threshold
  */
 export const MAX_QUEUE_THRESHOLD = 100;
 
 // === API MANAGEMENT ===
-
-/**
- * API request timeout in seconds
- * All ARR API calls use this consistent timeout
- */
-export const API_TIMEOUT_SECONDS = 120;
 
 /**
  * API usage thresholds for visual indicators (percentage)
@@ -75,11 +64,6 @@ export const API_USAGE_DANGER_THRESHOLD = 90;
  * Default number of days before an item can be re-searched.
  */
 export const DEFAULT_RESEARCH_AFTER_DAYS = 7;
-
-/**
- * Minimum re-search interval (1 day)
- */
-export const MIN_RESEARCH_AFTER_DAYS = 1;
 
 /**
  * Maximum re-search interval (90 days)
