+++ apps/web/src/features/trash-guides/components/wizard-steps/cf-configuration-types.ts
@@ -20,3 +20,19 @@ export interface ConditionEditorTarget {
 	trashId: string;
 	format: ConditionEditorFormat;
 }
+
+/** CF resolution decision when merging instance CFs with TRaSH template */
+export type CFResolutionDecision = "use_trash" | "keep_instance";
+
+/** Resolved CF data to be passed forward from the cf-resolution step */
+export interface ResolvedCF {
+	instanceCFId: number;
+	instanceCFName: string;
+	decision: CFResolutionDecision;
+	/** If decision is use_trash, this contains the TRaSH CF trash_id */
+	trashId?: string;
+	/** Recommended score from TRaSH if applicable */
+	recommendedScore?: number;
+	/** Original instance score */
+	instanceScore?: number;
+}
