+++ apps/web/src/lib/api-client/auth/index.ts
@@ -18,7 +18,6 @@ export type {
 	CurrentUser,
 	DeviceType,
 	OIDCLoginResponse,
-	OIDCProvider,
 	OIDCProviderResponse,
 	PasskeyCredential,
 	PasskeyCredentialsResponse,
