+++ package.json
@@ -40,6 +40,7 @@
     "@playwright/test": "^1.58.2",
     "dotenv": "^17.3.1",
     "eslint-plugin-unused-imports": "^4.4.1",
+    "knip": "^6.1.0",
     "turbo": "^2.9.1",
     "typescript": "^5.9.3"
   }
